+++ CHANGELOG.rst
@@ -1,8 +1,16 @@
 Changelog
 =========
 
+`Unreleased`_
+-------------
+
+- added filter to allow search directly in private 1:1 chats.
+- improved command help.
+
 1.0.0
 -----
 
 - initial release
 
+
+.. _Unreleased: https://github.com/simplebot-org/simplebot_lyrics/compare/v1.0.0...HEAD